1.1-beta2 comes with a statistic sidebar widget
After adding some stats to the right now bbPress dashboard widget, you
can now add a sidebar widget that shows global statistics or forum's
statistics if on a forum page. Clicking on the links of the widget
filters the topics to only show the desired support status (Resolved or
Not Resolved).
Sidebar widget is only available for bbPress 2.2.3 and above.